# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Child {
  id: number;
  name: string;
  age: number;
  interests: string[];
  deviceName: string;
  screenTimeWeekday: number;
  screenTimeWeekend: number;
  appleArcade: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface ChildInput {
  name: string;
  age: number;
  interests: string[];
  deviceName: string;
  screenTimeWeekday: number;
  screenTimeWeekend: number;
  appleArcade: boolean;
}

export interface CatalogApp {
  id: number;
  name: string;
  appStoreUrl: string;
  category: string;
  ageMin: number;
  ageMax: number;
  interestTags: string[];
  costModel: string;
  adStatus: string;
  notes: string;
  lastVerified: string;
  status: string;
  createdAt: string;
  updatedAt: string;
}

export interface AppInput {
  name: string;
  appStoreUrl: string;
  category: string;
  ageMin: number;
  ageMax: number;
  interestTags: string[];
  costModel: string;
  adStatus: string;
  notes?: string;
  status?: string;
}

export interface ChildAppStatusRecord {
  id: number;
  childId: number;
  appId: number;
  status: string;
  createdAt: string;
  updatedAt: string;
}

export interface ChildAppStatusInput {
  status: string;
}

export interface ExportData {
  children: Child[];
  apps: CatalogApp[];
  childAppStatuses: ChildAppStatusRecord[];
}

export type ListAppsParams = {
includeRemoved?: boolean;
category?: string;
interestTag?: string;
};