# lib/api-zod/src/generated/types/appInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AppInput {
  name: string;
  appStoreUrl: string;
  category: string;
  ageMin: number;
  ageMax: number;
  interestTags: string[];
  costModel: string;
  adStatus: string;
  notes?: string;
  status?: string;
}