# lib/api-zod/src/generated/types/child.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Child {
  id: number;
  name: string;
  age: number;
  interests: string[];
  deviceName: string;
  screenTimeWeekday: number;
  screenTimeWeekend: number;
  appleArcade: boolean;
  createdAt: string;
  updatedAt: string;
}