# lib/api-zod/src/generated/types/childAppStatusRecord.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ChildAppStatusRecord {
  id: number;
  childId: number;
  appId: number;
  status: string;
  createdAt: string;
  updatedAt: string;
}