# lib/api-zod/src/generated/types/childInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ChildInput {
  name: string;
  age: number;
  interests: string[];
  deviceName: string;
  screenTimeWeekday: number;
  screenTimeWeekend: number;
  appleArcade: boolean;
}