# lib/api-zod/src/generated/types/exportData.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CatalogApp } from './catalogApp';
import type { Child } from './child';
import type { ChildAppStatusRecord } from './childAppStatusRecord';

export interface ExportData {
  children: Child[];
  apps: CatalogApp[];
  childAppStatuses: ChildAppStatusRecord[];
}